Implement a mock driver
It provides a stub implementation of the harness.Driver and harness.Device interfaces.
And utilizes VT100 emulation[1] to parse and emulate a console for createAnsibleInventory to work.

It stores persistent data (name/tags) in a json file hardcoded at `/tmp/jumpstarter-mock.json`

The mock driver is not enabled by default and can be built with `go build -tags mock`

[1] https://pkg.go.dev/golang.org/x/term