Improve Lease Release & Exporter handling
There is a corner case, if we update a lease as Ended, and then we try
to update the Exporter but we fail (i.e. API becomes unavailable), the
exporter will remain with a "Lease" in state.

This commit first attempts to remove the lease from the exporter,
double checking that the lease is ours, then updates the lease as ended.