Dutlink driver: fix timing issues with power sequences
Power on/off for dutlink can be as simple as controlling power, but
when power sequences are configured in the hardware it could be
as involved as ... lowering a pin, waiting 10s, raising the pin.. etc,
which in the current firmware design can leave the usb loop blocked
for some time.

Raising the timeout fixes the timeout issues.