give exporters some time to disconnect
this is supposed to help with a flaky test "can lease and connect to exporters". The delete leases operation before this test causes exporters to disconnect, but it can take a short while to initiate that, and the kubectl wait in wait_for_exporter() can catch the Online/Registered state from before the disconnect. Let's see if it helps...