Offline exporters should be Pending and not Unsatisfiable
There are situations where an exporter is the only existing one, but
the service could be restarting, or the exporter could be rebooting
due to OS updates. We don't want to make it unsatisfiable,
but instead keep trying until the node comes back.

We leave the decission of canceling the lease to the client,
which will be informed of the reason for which the lease is pending
"Offline"