server-side proxy calls
When using a ref: that is internally used by another driver on the server side, forward the calls to the target driver.
Needs to resolve the target device first, and for that we need the actual root device as a reference point to start the search from.
Raise on improper use - trying to access report() or other driver call before the target is resolved.
Also change report() to not require "root", as the function always returns a report of itself - or the proxy target in case of Proxy driver

(cherry picked from commit ba567ee50e0727adfba7c8d97d642460efd2eb98)